Update hotspot share to app link url.
